refactor: common으로 Section 컴포넌트
