+++ packages/web-domains/src/result/features/common/components/index.ts
@@ -1,0 +1,1 @@
+export { Section } from './Section/Section';
