+++ packages/web-domains/src/result/features/main/components/index.ts
@@ -1,5 +1,4 @@
 export { HeaderBanner } from './HeaderBanner/HeaderBanner';
-export { SelectedDataCard } from './SelectedDataCard/SelectedDataCard';
 export { CountByMemberList } from './SelectedDataCard/CountByMemberList';
 export { MostAnswered, WithMeMembers } from './SelectedDataCard/Content';
 export { Comment } from './SelectedDataCard/Comment';
