+++ packages/web-domains/src/result/features/main/containers/CommentListContainer.tsx
@@ -1,13 +1,14 @@
 import { size } from '@sambad/sds/theme';
-import { Comment, SelectedDataCard } from '../components';
+import { Comment } from '../components';
+import { Section } from '../../common/components';
 
 export const CommentListContainer = () => {
   return (
-    <SelectedDataCard title="모임원들의 코멘트" style={{ marginTop: size['2xs'] }}>
+    <Section title="모임원들의 코멘트" style={{ marginTop: size['2xs'] }}>
       <Comment name="이세민" imgUrl="" comment="헬로 월드" />
       <Comment name="이세민" imgUrl="" comment="헬로 월드" />
       <Comment name="이세민" imgUrl="" comment="헬로 월드" />
       <Comment name="이세민" imgUrl="" comment="헬로 월드" />
-    </SelectedDataCard>
+    </Section>
   );
 };
