+++ packages/web-domains/src/result/features/main/containers/WithMyMembersContainer.tsx
@@ -1,9 +1,10 @@
 import { size } from '@sambad/sds/theme';
-import { SelectedDataCard, WithMeMembers } from '../components';
+import { WithMeMembers } from '../components';
+import { Section } from '../../common/components';
 
 export const WithMyMembersContainers = () => {
   return (
-    <SelectedDataCard title="나와 같이 축구, 농구, 수영을 선택한 모임원은?" style={{ marginTop: size['2xs'] }}>
+    <Section title="나와 같이 축구, 농구, 수영을 선택한 모임원은?" style={{ marginTop: size['2xs'] }}>
       <WithMeMembers
         count={5}
         members={[
@@ -11,6 +12,6 @@ export const WithMyMembersContainers = () => {
           { name: '이세민', imgUrl: '', id: 1 },
         ]}
       />
-    </SelectedDataCard>
+    </Section>
   );
 };
